+++ packages/webdriver/src/bidi/core.ts
@@ -18,7 +18,14 @@ export class BidiCore {
         this.#ws = new Socket(this._webSocketUrl, opts) as WebSocket
     }
 
-    public connect () {
+    public async connect () {
+        /**
+         * don't connect and stale unit tests when the websocket url is set to a dummy value
+         * Note: the value is defined in __mocks__/fetch.ts
+         */
+        if (process.env.VITEST_WORKER_ID && this._webSocketUrl === 'ws://webdriver.io') {
+            return
+        }
         return new Promise<void>((resolve) => this.#ws.on('open', () => {
             log.info('Connected session to Bidi protocol')
             this.#isConnected = true
