+++ packages/webdriverio/src/commands/browser/throttle.ts
@@ -1,4 +1,4 @@
-import type { ThrottleOptions } from '../../utils/interception/types.js'
+import type { ThrottleOptions } from '../../types.js'
 import logger from '@wdio/logger'
 import { getBrowserObject } from '@wdio/utils'
