+++ packages/webdriverio/tests/commands/browser/mock.test.ts
@@ -1,102 +1,34 @@
 import path from 'node:path'
-import { expect, describe, it, beforeEach, vi } from 'vitest'
+import { expect, describe, it, vi } from 'vitest'
 
 import { remote } from '../../../src/index.js'
+import { SESSION_MOCKS } from '../../../src/commands/browser/mock.js'
 
-vi.mock('fetch')
 vi.mock('@wdio/logger', () => import(path.join(process.cwd(), '__mocks__', '@wdio/logger')))
-vi.mock('@wdio/utils', async (origMod) => {
-    const orig: any = await origMod()
-    return {
-        ...orig,
-        userImport: vi.fn().mockResolvedValue({})
-    }
-})
-
-const clientMock: any = {
-    send: vi.fn(),
-    on: vi.fn()
-}
-const pageMock: any = {
-    target: vi.fn().mockReturnValue({
-        createCDPSession: vi.fn().mockReturnValue(Promise.resolve(clientMock))
-    }),
-    evaluate: vi.fn().mockReturnValue(Promise.resolve(true))
-}
-const puppeteerMock: any = {
-    pages: vi.fn().mockReturnValue([pageMock]),
-    isConnected: vi.fn().mockReturnValue(true)
-}
-
-vi.mock('../../../src/utils/interception/webdriver', () => ({
-    default: class {
-        init = vi.fn()
-    }
-}))
 
 describe('mock', () => {
     let browser: WebdriverIO.Browser
-    beforeEach(async () => {
-        clientMock.send.mockClear()
-    })
 
-    it('should enable the fetch domain if not already done', async () => {
+    it('should throw if Bidi is not used', async () => {
         browser = await remote({
             baseUrl: 'http://foobar.com',
             capabilities: {
-                browserName: 'devtools'
+                browserName: 'chrome'
             }
         })
 
-        browser.puppeteer = puppeteerMock
-
-        expect(clientMock.send).toBeCalledTimes(0)
-        await browser.mock('/foobar')
-        expect(clientMock.send).toBeCalledWith('Fetch.enable', expect.any(Object))
-        expect(clientMock.on).toBeCalledWith('Fetch.requestPaused', expect.any(Function))
-
-        await browser.mock('/foobar')
-        expect(clientMock.send).toBeCalledTimes(1)
+        await expect(() => browser.mock('')).rejects.toThrow(/only supported/)
     })
 
-    it('should mock on multiple tabs', async () => {
+    it('can initiate a mock', async () => {
         browser = await remote({
             baseUrl: 'http://foobar.com',
             capabilities: {
-                browserName: 'devtools'
+                browserName: 'bidi'
             }
         })
 
-        browser.puppeteer = puppeteerMock
-
-        // @ts-expect-error mock feature
-        vi.mocked(fetch).setMockResponse('window-handle-2')
-        expect(clientMock.send).toBeCalledTimes(0)
-        await browser.mock('/foobar')
-        expect(clientMock.send).toBeCalledTimes(1)
-        expect(clientMock.send).toBeCalledWith('Fetch.enable', expect.any(Object))
-        expect(clientMock.on).toBeCalledWith('Fetch.requestPaused', expect.any(Function))
-        // @ts-expect-error mock feature
-        vi.mocked(fetch).setMockResponse('window-handle-3')
-        await browser.mock('/foobar')
-        expect(clientMock.send).toBeCalledTimes(2)
-
-    })
-
-    it('should mock using WebDriver capabilities', async () => {
-        browser = await remote({
-            baseUrl: 'http://foobar.com',
-            capabilities: {
-                browserName: 'chrome',
-                'sauce:options': {
-                    extendedDebugging: true
-                }
-            }
-        })
-
-        browser.puppeteer = puppeteerMock
-        const mock = await browser.mock('/foobar')
-        // @ts-expect-error mock feature
-        expect(vi.mocked(mock.init)).toBeCalledWith()
+        const mock = browser.mock('**/*').catch(() => { /* ignore */ })
+        expect(SESSION_MOCKS[browser.sessionId]).toContain(mock)
     })
 })
